+++ fotoskazka/app/Filament/Resources/Icons/Pages/CreateIcon.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\Icons\Pages;
+
+use App\Filament\Resources\Icons\IconResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateIcon extends CreateRecord
+{
+    protected static string $resource = IconResource::class;
+}